Add files via upload
Separated UPDI for type1 (bitbang) and type3 (UART). Changed to more 'relaxed' timings for using UPDI.
