+++ source/UPDI_lo_lvl.h
@@ -9,9 +9,13 @@
 #ifndef UPDI_LO_LVL_H_
 #define UPDI_LO_LVL_H_
 
-#include <stdint.h>
 #include "sys.h"
-#include "updi_io.h"
+#if UPDI_IO_TYPE==3
+  #include "updi_io_uart.h"
+#else
+  #include "updi_io.h"
+#endif
+//#include "updi_io.h"
 
 namespace UPDI {
 	// UPDI registers
